opzet controllers
Basis controller classes gemaakt. Later specifieker maken.

+++ src/main/java/com/woningzoeker/woningzoeker/repositories/ContactInfoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ContactInfoRepository extends JpaRepository<ContactInfo, Integer> {
+public interface ContactInfoRepository extends JpaRepository<ContactInfo, Long> {
     List<ContactInfo> findByProfielId(long profielId);
     List<ContactInfo> findByGebruikerId(long gebruikerId);
 }
