+++ src/main/java/com/woningzoeker/woningzoeker/repositories/HuisRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface HuisRepository extends JpaRepository<Huis, Integer> {
+public interface HuisRepository extends JpaRepository<Huis, Long> {
     List<Huis> findByAdres(String Adres);
 
 }
