+++ src/main/java/com/woningzoeker/woningzoeker/repositories/LocatieRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface LocatieRepository extends JpaRepository<Locatie, Integer> {
+public interface LocatieRepository extends JpaRepository<Locatie, Long> {
     List<Locatie> findByWoonplaats(String Woonplaats);
 }
